Abstracted foreign functions and moved standard library to its own project
