+++ Outlet.ForeignFunctions/ForeignConstructor.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Outlet.ForeignFunctions
+{
+    [AttributeUsage(AttributeTargets.Constructor, AllowMultiple = true)]
+    public sealed class ForeignConstructor : Attribute
+    {
+    }
+}
