+++ Outlet/Outlet.csproj
@@ -23,4 +23,9 @@
     <PackageReference Include="NETStandard.Library" Version="2.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Outlet.ForeignFunctions\Outlet.ForeignFunctions.csproj" />
+    <ProjectReference Include="..\Outlet.StandardLib\Outlet.StandardLib.csproj" />
+  </ItemGroup>
+
 </Project>
